+++ internal/models/volume/Volume.go
@@ -11,7 +11,7 @@ type Volume struct {
 	ID                primitive.ObjectID `bson:"_id" json:"id,omitempty"`
 	StockId           primitive.ObjectID `bson:"_stock_id" json:"stock_id,omitempty"`
 	Date              time.Time          `json:"date"`
-	ShortVolume       uint64             `json:"shortVolume"`
-	ShortExemptVolume uint64             `json:"shortExemptVolume"`
-	TotalVolume       uint64             `json:"totalVolume"`
+	ShortVolume       float64            `json:"shortVolume"`
+	ShortExemptVolume float64            `json:"shortExemptVolume"`
+	TotalVolume       float64            `json:"totalVolume"`
 }
